refactored proposals page
