+++ src/components/proposals/CreateProposalsView.tsx
