+++ src/components/proposals/ProposalsView.tsx
