+++ src/components/proposals/components/ProposalCard.tsx
@@ -3,16 +3,21 @@ import relativeTime from "dayjs/plugin/relativeTime";
 import dayjs from "dayjs";
 import { truncateAddress } from "@utils/addresses";
 import { upperCase } from "lodash";
-import { TokenVotingProposalListItem } from "@aragon/sdk-client";
+import { type TokenVotingProposalListItem } from "@aragon/sdk-client";
+import { useRouter } from "next/router";
 
 dayjs.extend(relativeTime);
 
 type Props = TokenVotingProposalListItem;
 
 export const ProposalCard: React.FC<Props> = ({ status, ...proposal }) => {
+  const router = useRouter();
   return (
     <div className="card col-span-1 w-full bg-base-100 shadow-xl">
-      <div className="card-body">
+      <div
+        className="card-body"
+        onClick={() => router.push(`/proposals/${proposal.id}`)}
+      >
         <div className="card-actions justify-between">
           <div className="flex flex-1 items-center justify-start gap-2 text-sm font-medium">
             <div className="placeholder avatar">
@@ -23,7 +28,8 @@ export const ProposalCard: React.FC<Props> = ({ status, ...proposal }) => {
               </div>
             </div>
             <p className="text-md">
-              {proposal.dao.name} by <strong>{proposal.dao.address}</strong>
+              {proposal.dao.name} by{" "}
+              <strong>{truncateAddress(proposal.dao.address, 8, 6)}</strong>
             </p>
           </div>
